Restructure files and move NavBar one level up
