+++ client/src/actions/types.js
@@ -9,7 +9,7 @@ export const FETCH_FOOS = 'FETCH_FOOS';
 export const GET_PUBLICATION_BY_ID = 'GET_PUBLICATION_BY_ID';
 export const GET_PUBLICATIONS_BY_TEAM_ID = 'GET_PUBLICATIONS_BY_TEAM_ID';
 
-export const ADD_USER_DATA = 'ADD_USER_DATE';
+export const ADD_USER_DATA = 'ADD_USER_DATA';
 // export const CREATE_PUBLICATION = 'CREATE_PUBLICATION'
 // export const UPDATE_PUBLICATION = 'UPDATE_PUBLICATION'
 // export const DELETE_PUBLICATION = 'DELETE_PUBLICATION'
