+++ client/src/api/api.js
@@ -1,0 +1,9 @@
+import axios from 'axios';
+
+
+ const usingRp = true;
+ const localServerUrl = usingRp ? '/api' : 'http://localhost:5000';
+ 
+ export default axios.create({
+     baseURL: localServerUrl
+ });
