+++ client/src/components/auth/Register.js
@@ -5,12 +5,11 @@
  import React from 'react';
  import './Register.css';
  import RegistrationForm from './RegistrationForm';
- import NavigationBar from '../auth/NavigationBar';
+ import NavigationBar from '../NavigationBar';
  
  export default function Register() {
     return(
         <div>
-            <NavigationBar/>
             <div id='registration-page'>
                 <RegistrationForm/>
             </div>
